Use `bedrock-tokenizer` to provide tokenizers.
