+++ lib/config.js
@@ -4,7 +4,6 @@
 import {config} from 'bedrock';
 
 config.tokenization = {};
-config.tokenization.kmsModule = 'ssm-v1';
 config.tokenization.tokenCreationConcurrency = 5;
 config.tokenization.defaultVersionOptions = {
   // sizes are in bytes
