+++ lib/documents.js
@@ -3,14 +3,16 @@
  */
 import * as bedrock from 'bedrock';
 import * as database from 'bedrock-mongodb';
+import {tokenizers} from 'bedrock-tokenizer';
 import {IdGenerator, IdEncoder} from 'bnid';
 import canonicalize from 'canonicalize';
 import {Cipher} from 'minimal-cipher';
 import {promisify} from 'util';
 import {createKeyResolver} from './keyResolver.js';
-import {hmacString} from './kms.js';
-import * as tokenizers from './tokenizers.js';
+import {TextEncoder} from 'util';
 const {util: {BedrockError}} = bedrock;
+
+const TEXT_ENCODER = new TextEncoder();
 
 // 128 bit random id generator
 const idGenerator = new IdGenerator({bitLength: 128});
@@ -100,7 +102,7 @@ export async function register({
   // threat of a stolen database. Once HMAC'd, dictionary attacks may be more
   // difficult -- particularly if the HMAC material is in an HSM.
   const [externalIdHash, hash] = await Promise.all([
-    hmacString({hmac, value: externalId}),
+    _hmacString({hmac, value: externalId}),
     _hmacDocument({hmac, document})
   ]);
 
@@ -245,5 +247,13 @@ async function _refresh({externalIdHash, hash, ttl, registrationData}) {
 async function _hmacDocument({hmac, document}) {
   // ensure document is in canonical form before hashing
   const value = canonicalize(document);
-  return hmacString({hmac, value});
+  return _hmacString({hmac, value});
+}
+
+async function _hmacString({hmac, value}) {
+  const data = TEXT_ENCODER.encode(value);
+  const signature = await hmac.sign({data});
+  // TODO: worth using hmac type to add prefix to signature for future proofing?
+  // ... `sha256:<signature>`? ... note that this could complicate key rotation
+  return signature;
 }
