+++ lib/main.js
@@ -5,7 +5,6 @@ import './config.js';
 
 import * as documents from './documents.js';
 import * as tokens from './tokens.js';
-import * as tokenizers from './tokenizers.js';
 import * as tokenVersions from './tokenVersions.js';
 
-export {documents, tokens, tokenizers, tokenVersions};
+export {documents, tokens, tokenVersions};
