+++ lib/tokens.js
@@ -4,13 +4,13 @@
 import * as base64url from 'base64url-universal';
 import * as bedrock from 'bedrock';
 import * as database from 'bedrock-mongodb';
+import {tokenizers} from 'bedrock-tokenizer';
 import {IdGenerator, IdEncoder} from 'bnid';
 import {randomBytes, timingSafeEqual} from 'crypto';
 import pLimit from 'p-limit';
 import {promisify} from 'util';
 import {createKek} from './aeskw.js';
 import Bitstring from './Bitstring.js';
-import * as tokenizers from './tokenizers.js';
 import * as tokenVersions from './tokenVersions.js';
 const randomBytesAsync = promisify(randomBytes);
 const {util: {BedrockError}} = bedrock;
@@ -58,9 +58,8 @@ bedrock.events.on('bedrock-mongodb.ready', async () => {
 });
 
 /**
- * Creates `tokenCount` many unique tokens using the given `tokenizer`, each
- * bound to the same given `internalId`, having the given authenticated but
- * cleartext `attributes`.
+ * Creates `tokenCount` many unique each bound to the same given `internalId`,
+ * having the given authenticated but cleartext `attributes`.
  *
  * To an external (outside of the system using this module) party, each token
  * has a random appearance, but internally (within the system using this module)
@@ -112,9 +111,8 @@ bedrock.events.on('bedrock-mongodb.ready', async () => {
  *
  * @returns {object} An object with `tokens` as an array of created tokens.
  */
-export async function create({
-  tokenizer, internalId, attributes = new Uint8Array(), tokenCount
-} = {}) {
+export async function create(
+  {internalId, attributes = new Uint8Array(), tokenCount} = {}) {
   // TODO: validate `tokenCount`
   if(!(attributes instanceof Uint8Array)) {
     throw new TypeError('"attributes" must be a Uint8Array.');
@@ -123,7 +121,8 @@ export async function create({
     throw new TypeError('"attributes" maximum size is 14 bytes.');
   }
 
-  const {id: tokenizerId, hmac} = tokenizer;
+  // get the current tokenizer and its HMAC API
+  const {id: tokenizerId, hmac} = await tokenizers.getCurrent();
 
   // get version associated with tokenizer, creating it as needed
   const {tokenVersion} = await tokenVersions.ensureTokenVersion(
