+++ package.json
@@ -32,13 +32,13 @@
     "isomorphic-webcrypto": "^2.3.6",
     "minimal-cipher": "^1.0.0",
     "p-limit": "^3.0.1",
-    "pako": "^1.0.11",
-    "webkms-client": "^2.1.0"
+    "pako": "^1.0.11"
   },
   "peerDependencies": {
     "bedrock": "3.x",
     "bedrock-https-agent": "^1.0.1",
-    "bedrock-mongodb": "7.x"
+    "bedrock-mongodb": "7.x",
+    "bedrock-tokenizer": "github:digitalbazaar/bedrock-tokenizer#initial"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
